+++ example/lib/liquid_circular_progress_indicator_page.dart
@@ -1,0 +1,101 @@
+import 'package:flutter/material.dart';
+import 'package:liquid_progress_indicator/liquid_progress_indicator.dart';
+
+class LiquidCircularProgressIndicatorPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Liquid Circular Progress Indicators"),
+      ),
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+        children: <Widget>[
+          _AnimatedLiquidCircularProgressIndicator(),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            children: <Widget>[
+              SizedBox(
+                width: 75,
+                height: 75,
+                child: LiquidCircularProgressIndicator(
+                  backgroundColor: Colors.black,
+                  valueColor: AlwaysStoppedAnimation(Colors.red),
+                ),
+              ),
+              SizedBox(
+                width: 75,
+                height: 75,
+                child: LiquidCircularProgressIndicator(
+                  backgroundColor: Colors.white,
+                  valueColor: AlwaysStoppedAnimation(Colors.pink),
+                  borderColor: Colors.red,
+                  borderWidth: 5.0,
+                ),
+              ),
+              SizedBox(
+                width: 75,
+                height: 75,
+                child: LiquidCircularProgressIndicator(
+                  backgroundColor: Colors.white,
+                  valueColor: AlwaysStoppedAnimation(Colors.grey),
+                  borderColor: Colors.blue,
+                  borderWidth: 5.0,
+                  center: Icon(Icons.cancel),
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class _AnimatedLiquidCircularProgressIndicator extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() =>
+      _AnimatedLiquidCircularProgressIndicatorState();
+}
+
+class _AnimatedLiquidCircularProgressIndicatorState
+    extends State<_AnimatedLiquidCircularProgressIndicator>
+    with SingleTickerProviderStateMixin {
+  AnimationController _animationController;
+
+  @override
+  void initState() {
+    super.initState();
+    _animationController = AnimationController(
+      vsync: this,
+      duration: Duration(seconds: 10),
+    );
+
+    _animationController.addListener(() => setState(() {}));
+    _animationController.repeat();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final percentage = _animationController.value * 100;
+    return Center(
+      child: SizedBox(
+        width: 150.0,
+        height: 150.0,
+        child: LiquidCircularProgressIndicator(
+          value: _animationController.value,
+          backgroundColor: Colors.white,
+          valueColor: AlwaysStoppedAnimation(Colors.blue),
+          center: Text(
+            "${percentage.toStringAsFixed(0)}%",
+            style: TextStyle(
+              color: Colors.lightBlueAccent,
+              fontSize: 20.0,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}