+++ example/lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
-import 'home.dart';
+import 'liquid_circular_progress_indicator_page.dart';
 
-void main() => runApp(MaterialApp(home: Home()));
+void main() => runApp(MaterialApp(
+      home: LiquidCircularProgressIndicatorPage(),
+    ));