+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.0]
+
+* Added borderColor and borderWidth properties to LiquidCircularProgressIndicator.
+* Added more examples to the example project.
+
 ## [0.0.2]
 
 * Updated documentation.